Add javascript.js and write three functions
i wrote three functions in the javascript. The first function is
computerPlay to randomly select from three options, the second one is
playerSelection to let player choose what hand to play, and the last one
is singleRound to compare the computer hand .vs player hand and announce
if player had lose or win the game in that round.