Edit javascript.js and add a new function
i add separate score variable to store the score of player and computer,
and I edit the singleRound function to return more informative string
and add one score for the winner of the round. i also add a new function
game() to loop the singleRound five time and announce who the winner is
when the loop ended.